pyproject.toml
[project]
name = "my-hr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.2.0",
]
